+++ src/platform/site-wide/user-nav/components/SearchMenu.jsx
