+++ src/platform/site-wide/user-nav/components/SignInProfileMenu.jsx
