+++ src/platform/site-wide/user-nav/tests/components/SearchHelpSignIn.unit.spec.jsx
