+++ src/platform/site-wide/user-nav/tests/components/SearchMenu.unit.spec.jsx
