+++ src/platform/site-wide/user-nav/tests/user-nav.e2e.spec.js
